Wrap containers for async import
