+++ x-pack/plugins/cases/public/components/visualizations/actions/add_to_existing_case.tsx
@@ -4,53 +4,23 @@
  * 2.0; you may not use this file except in compliance with the Elastic License
  * 2.0.
  */
-import React, { useEffect, useMemo } from 'react';
+import React from 'react';
 import { unmountComponentAtNode } from 'react-dom';
-import type { Embeddable as LensEmbeddable } from '@kbn/lens-plugin/public';
 import { createAction } from '@kbn/ui-actions-plugin/public';
 import { isErrorEmbeddable } from '@kbn/embeddable-plugin/public';
 
 import { toMountPoint } from '@kbn/kibana-react-plugin/public';
 
 import type { CaseUI } from '../../../../common';
-import { isLensEmbeddable, hasInput, getLensCaseAttachment } from './utils';
+import { isLensEmbeddable, hasInput /* getLensCaseAttachment*/ } from './utils';
 
 import type { ActionContext, CasesUIActionProps } from './types';
-import { useCasesAddToExistingCaseModal } from '../../all_cases/selector_modal/use_cases_add_to_existing_case_modal';
 import { ADD_TO_EXISTING_CASE_DISPLAYNAME } from './translations';
-import { ActionWrapper } from './action_wrapper';
 import { canUseCases } from '../../../client/helpers/can_use_cases';
 import { getCaseOwnerByAppId } from '../../../../common/utils/owner';
 
 export const ACTION_ID = 'embeddable_addToExistingCase';
 export const DEFAULT_DARK_MODE = 'theme:darkMode' as const;
-
-interface Props {
-  embeddable: LensEmbeddable;
-  onSuccess: () => void;
-  onClose: (theCase?: CaseUI) => void;
-}
-
-const AddExistingCaseModalWrapper: React.FC<Props> = ({ embeddable, onClose, onSuccess }) => {
-  const modal = useCasesAddToExistingCaseModal({
-    onClose,
-    onSuccess,
-  });
-
-  const attachments = useMemo(() => {
-    const { timeRange } = embeddable.getInput();
-    const attributes = embeddable.getFullAttributes();
-    // we've checked attributes exists before rendering (isCompatible), attributes should not be undefined here
-    return attributes != null ? [getLensCaseAttachment({ attributes, timeRange })] : [];
-  }, [embeddable]);
-  useEffect(() => {
-    modal.open({ getAttachments: () => attachments });
-  }, [attachments, modal]);
-
-  return null;
-};
-
-AddExistingCaseModalWrapper.displayName = 'AddExistingCaseModalWrapper';
 
 export const createAddToExistingCaseLensAction = ({
   core,
@@ -88,6 +58,7 @@ export const createAddToExistingCaseLensAction = ({
     },
     execute: async ({ embeddable }) => {
       const targetDomElement = document.createElement('div');
+      const Wrapper = React.lazy(() => import('./wrappers/add_to_existing_case_wrapper'));
 
       const cleanupDom = (shouldCleanup?: boolean) => {
         if (targetDomElement != null && shouldCleanup) {
@@ -109,21 +80,11 @@ export const createAddToExistingCaseLensAction = ({
       const onSuccess = () => {
         cleanupDom(true);
       };
+
       const mount = toMountPoint(
-        <ActionWrapper
-          core={core}
-          caseContextProps={caseContextProps}
-          storage={storage}
-          plugins={plugins}
-          history={history}
-          currentAppId={currentAppId}
-        >
-          <AddExistingCaseModalWrapper
-            embeddable={embeddable}
-            onClose={onClose}
-            onSuccess={onSuccess}
-          />
-        </ActionWrapper>,
+        <Wrapper
+          {...{ embeddable, onClose, onSuccess, core, plugins, storage, history, caseContextProps }}
+        />,
         { theme$: theme.theme$ }
       );
