+++ x-pack/plugins/cases/public/containers/toaster_error.ts
@@ -1,0 +1,18 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export const createToasterPlainError = (message: string) => new ToasterError([message]);
+
+export class ToasterError extends Error {
+  public readonly messages: string[];
+
+  constructor(messages: string[]) {
+    super(messages[0]);
+    this.name = 'ToasterError';
+    this.messages = messages;
+  }
+}
